# backend/open_api_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "budget-tool_backend",
    "version": "0.2.0"
  },
  "paths": {
    "/": {
      "get": {
        "operationId": "app_version",
        "responses": {
          "200": {
            "description": "Version of the server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HomeResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to get the vesion of the server",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/auth/validate-token": {
      "get": {
        "summary": "Validate the JWT token in the header.\nUm I kinda forgot to implement this route. So this route will always return true",
        "operationId": "validate_token",
        "responses": {
          "200": {
            "description": "The JWT token is still valid"
          },
          "401": {
            "description": "Authenication token expired. Please reauthenicate",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Authenication failed",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Failed to get update spending item",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/auth/validate-token/v2": {
      "get": {
        "summary": "Validate the JWT token in the header fr this time.\nBy checking the JWT token in the authorization header is still valid or not",
        "operationId": "validate_token_v2",
        "responses": {
          "200": {
            "description": "The JWT token is still valid"
          },
          "403": {
            "description": "JWT has expired",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/budget/{year}/{month}": {
      "get": {
        "summary": "Get the budget information for a specific month",
        "operationId": "get_month_budget_handler",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "description": "The year which to get the budget of",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "month",
            "in": "path",
            "description": "The month's budget to get. The first letter of the month's name is expected to the captalized. ie `January`",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Month"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The requested month's budget. If the request month does not have any budget records, this route will iterate back till either no more months to check, a budget is encountered. The returned budget will have no spending, all the fields are correct, and matches the request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonthlyBudget"
                }
              }
            }
          },
          "401": {
            "description": "Authenication token expired. Please reauthenicate",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Authenication failed",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "All months, the requested month and before does not contain any monthly budget",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Failed to get the requested month's budget",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "deprecated": true
      },
      "post": {
        "summary": "Update the budget for a specific month in a specific year. This route will also ensure the `totalSpending`, and `overBudgetAmount` is up to date",
        "operationId": "update_budget_handler",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "description": "The year which to update the budget of",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "month",
            "in": "path",
            "description": "The month's budget to update. The first letter of the month's name is expected to the captalized. ie `January`",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Month"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MonthlyBudget"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully updated the month's budget"
          },
          "401": {
            "description": "Authenication token expired. Please reauthenicate",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Authenication failed",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Failed to update the month's budget",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "deprecated": true
      }
    },
    "/export/{year}/{month}/csv": {
      "get": {
        "summary": "Gets the spending items in csv form. If converting budget information to CSV fails at any point, partial data will not be returned",
        "operationId": "export_csv_handler",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "description": "The year the spending item is in",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "month",
            "in": "path",
            "description": "The month the spending item is in",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Month"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Spending items in CSV format",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Authenication token expired. Please reauthenicate",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Authenication failed",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Failed to get update spending item",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/login/basic": {
      "post": {
        "operationId": "basic_auth_handler",
        "responses": {
          "200": {
            "description": "Login successful. Returns a JWT token",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Authenication token expired. Please reauthenicate",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Authenication failed",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/login/basic/v2": {
      "post": {
        "summary": "Better implemented version of the basic auth handler, using proper axum_extra constructs. This handler should behave the exact same as previous",
        "operationId": "basic_auth_handler_v2",
        "responses": {
          "200": {
            "description": "Login successful. Returns a JWT token",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Authenication token expired. Please reauthenicate",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Authenication failed",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/notification/save-subscription": {
      "post": {
        "operationId": "save_notification_subscription_handler",
        "requestBody": {
          "description": "The notfication subscription generated by the browser",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationSubscription"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Notification sent successfully"
          }
        },
        "deprecated": true
      }
    },
    "/notification/send": {
      "post": {
        "summary": "Send a notification",
        "operationId": "send_notification_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationSendBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Notification sent successfully"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HomeResponse": {
        "type": "object",
        "required": [
          "version"
        ],
        "properties": {
          "version": {
            "type": "string"
          }
        }
      },
      "Month": {
        "type": "string",
        "enum": [
          "January",
          "February",
          "March",
          "April",
          "May",
          "June",
          "July",
          "August",
          "September",
          "October",
          "November",
          "December"
        ]
      },
      "MonthlyBudget": {
        "type": "object",
        "description": "Budget details for single month",
        "required": [
          "month",
          "budget",
          "totalSpending",
          "overBudgetAmount",
          "spending"
        ],
        "properties": {
          "budget": {
            "$ref": "#/components/schemas/BudgetConfig",
            "description": "Budget details"
          },
          "carriedOverFrom": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Month",
                "description": "The month it was carried over from\nIf the setting are not carried over from a previous month, this value will be empty"
              }
            ]
          },
          "month": {
            "$ref": "#/components/schemas/Month",
            "description": "The month"
          },
          "overBudgetAmount": {
            "type": "number",
            "format": "double",
            "description": "Amount over budget for the month. 0 means not over budget."
          },
          "spending": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpendingItem"
            },
            "description": "List of spent items"
          },
          "totalSpending": {
            "type": "number",
            "format": "double",
            "description": "Total spending for the month. Including any over budget amount"
          }
        }
      },
      "NotificationBody": {
        "type": "object",
        "required": [
          "title",
          "body"
        ],
        "properties": {
          "body": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "NotificationKeys": {
        "type": "object",
        "required": [
          "p256dh",
          "auth"
        ],
        "properties": {
          "auth": {
            "type": "string"
          },
          "p256dh": {
            "type": "string"
          }
        }
      },
      "NotificationSendBody": {
        "type": "object",
        "required": [
          "subscription",
          "body"
        ],
        "properties": {
          "body": {
            "$ref": "#/components/schemas/NotificationBody",
            "description": "The body of the notification"
          },
          "subscription": {
            "$ref": "#/components/schemas/NotificationSubscription",
            "description": "Data sent by the browser to do the handshake for the notification"
          }
        }
      },
      "NotificationSubscription": {
        "type": "object",
        "description": "Stuff the browser sends to do the notification handshake",
        "required": [
          "endpoint",
          "keys"
        ],
        "properties": {
          "endpoint": {
            "type": "string"
          },
          "expirationTime": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          },
          "keys": {
            "$ref": "#/components/schemas/NotificationKeys"
          }
        }
      },
...
    }
  }
}